+++ ts/packages/memory/conversation/src/index.ts
@@ -1,4 +1,5 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+export * from "./podcast.js";
 export * from "./importPodcast.js";
